Minor: Fix more clang compiler warnings
Fix

```
../../../../ext/datadog_profiling_native_extension/collectors_thread_context.c:1830:37: error: implicit conversion loses integer precision: 'long' to 'int' [-Werror,-Wshorten-64-to-32]
 1830 |   for (int i = active_context_index - 1; i >= 0; i--) {
      |            ~   ~~~~~~~~~~~~~~~~~~~~~^~~
```

+++ ext/datadog_profiling_native_extension/collectors_thread_context.c
@@ -1827,7 +1827,7 @@ static void otel_without_ddtrace_trace_identifiers_for(
   otel_span local_root_span = active_span;
 
   // Now find the oldest span starting from the active span that still has the same trace id as the active span
-  for (int i = active_context_index - 1; i >= 0; i--) {
+  for (long i = active_context_index - 1; i >= 0; i--) {
     otel_span checking_span = otel_span_from(rb_ary_entry(context_storage, i), otel_current_span_key);
     if (checking_span.span == Qnil) return;
